+++ README.md
@@ -2,4 +2,4 @@
 
 Uses the Calendarific Holiday API endpoint to display holidays for a selected country for the current year. 
 
-Requires an API key from Calendarific. 
+Requires an API key from Calendarific.
