+++ src/components/CountrySearchField/SearchCountryInputField.tsx
@@ -1,14 +1,14 @@
-import React, {Dispatch, SetStateAction, useEffect} from "react";
-import {useState} from "react";
-import {supportedCountries, supportedCountryCodes, supportedCountryNames} from "../supportedCountries";
+import React, {Dispatch, SetStateAction, useEffect, useState} from "react";
+import {supportedCountries, supportedCountryCodes, supportedCountryNames} from "../../supportedCountries";
 import styled from "styled-components";
-
-type AutocompleteState = {
-    activeSuggestion: number,
-    filteredSuggestions: string[],
-    showSuggestions: boolean,
-    userInput: string
-}
+import {
+	InputFieldStyled,
+	NoSuggestionsStyled,
+	SearchCountryInputFieldWrapper,
+	StyledList,
+	StyledListItem
+} from "./SearchCountryInputField.styled";
+import {AutocompleteState} from "./SearchCountryInputField.types";
 
 export function SearchCountryInputField({ setSearchCountry}: { setSearchCountry: Dispatch<SetStateAction<{ code: string; isValid: boolean; }>> }) {
 	const [autocompleteState, setAutocompleteState] = useState<AutocompleteState>({
@@ -106,7 +106,7 @@ export function SearchCountryInputField({ setSearchCountry}: { setSearchCountry:
 	}
 
 	return (
-		<>
+		<SearchCountryInputFieldWrapper>
 			<InputFieldStyled
 				type="text"
 				placeholder="Start typing to search country..."
@@ -114,7 +114,7 @@ export function SearchCountryInputField({ setSearchCountry}: { setSearchCountry:
 				onKeyDown={onKeyDown}
 				value={autocompleteState.userInput}/>
 			{suggestionsListComponent}
-		</>
+		</SearchCountryInputFieldWrapper>
 	);
 }
 
@@ -128,50 +128,3 @@ function findCountryCode(input: string) {
 	return supportedCountries.find(c => c.name === input)!.iso.toLowerCase();
 }
 
-const StyledList = styled.ul`
-	border: 1px solid #c5cfdc;
-	border-top-width: 0;
-	border-radius: 0 0 5px 5px;
-	list-style: none;
-	margin-top: 3px;
-	max-height: 143px;
-	overflow-y: auto;
-	padding-left: 0;
-	width: 100%;
-	box-sizing: border-box;
-	font-size: 14px;
-	
-	position: absolute;
-	top: 38px;
-	left: 0;
-`;
-
-const StyledListItem = styled.li`
-	padding: 0.5rem;
-
-	&:hover {
-	 background-color: #f9fbfc;
-	 cursor: pointer;
-	 font-weight: 700;
- }
-	
-	&:not(:last-of-type) {
-		border-bottom: 1px solid #c5cfdc;
-	}
-`;
-
-const NoSuggestionsStyled = styled.div`
-	color: #c5cfdc;
-	font-size: 14px;
-	padding: 5px 0;
-`;
-
-export const InputFieldStyled = styled.input`
-	border-radius: 4px;
-	border: 1px solid hsl(0, 0%, 80%);
-	width: 300px;
-	height: 38px;
-	font-size: 14px;
-	padding-left: 10px;
-	box-sizing: border-box;
-`;
