+++ src/components/CountrySearchField/SearchCountryInputField.types.ts
@@ -1,0 +1,6 @@
+export type AutocompleteState = {
+    activeSuggestion: number,
+    filteredSuggestions: string[],
+    showSuggestions: boolean,
+    userInput: string
+}
