+++ src/components/SystemMessage/SystemMessage.styled.tsx
@@ -1,0 +1,19 @@
+import styled from "styled-components";
+
+export const SystemMessageWrapper = styled.div`
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    text-align: center;
+    margin: 200px 0;
+`;
+
+export const SystemMessageTitle = styled.div`
+    font-weight: 500;
+    font-size: 24px;
+    margin-bottom: 1rem;
+`;
+
+export const SystemMessageDescription = styled.div`
+    font-size: 14px;
+`;
