+++ src/components/SystemMessage/SystemMessage.types.ts
@@ -1,0 +1,5 @@
+export enum SystemMessageType {
+    EMPTY = "EMPTY",
+    API = "API",
+    ERROR = "ERROR",
+}
